requirements.txt
asgiref==3.8.1
Django==4.2.11
pillow==10.2.0
python-barcode==0.15.1
sqlparse==0.4.4
treepoem==3.24.0
typing_extensions==4.10.0
tzdata==2024.1
